feature-#85zrwzc7k/move MainScreen.tsx in new file and move App.tsx in MainScreen.tsx
